+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   Added optional `defaultMonth` prop to components `DatePickerField` and `DatePicker`.
+-   Added optional `imgProps` prop to component `Thumbnail`
+-   Added `isThumbnailCrossOriginEnabled` prop to `LinkPreview`.
+
+### Changed
+
+-   _[BREAKING]_ Replaced `today` and `monthOffset` props by `selectedMonth` for `DatePickerControlled` component.
+
 ### Fixed
 
 -   Trigger the infinite scroll callback on display if enough space with the `useInfiniteScroll` hook.